flasher: use avr_keyscanner firmware
Wow. That was a new sort of fun.

It flashed! I had to get the half-built firmware from the repo and then
get it building in this project (for dependencies) and then observed it
flashing each side's attiny88.

$ git clone https://github.com/Keyboardio/avr_keyscanner
$ nix-shell -p avrdude pkgsCross.avr.buildPackages.gcc python27
$ make all PRODUCT_ID=keyboardio-model-01

The flasher firmware sketch and the headers it uses were then copied to
this repository, built, and flashed to finish.

On boot, the device will flash each half (so make sure they're
connected). This can be seen by connecting to the serial console:

$ cu -9600 -l /dev/ttyACM*

Once it finishes both sides, then do as it says and flash the "real"
firmware back on and use it as before (again, it'll print a message to
the console).